Chore: Moved Utilities used by View to its own folder
