+++ EksamensProjekt/ViewModelsUtilities/ViewModelBase.cs
