+++ EksamensProjekt/Views/HistoryView.xaml
@@ -1,7 +1,7 @@
 ﻿<Window
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:util="clr-namespace:EksamensProjekt.Utilities"
+        xmlns:util="clr-namespace:EksamensProjekt.ViewsUtilities"
         xmlns:av="http://schemas.microsoft.com/expression/blend/2008" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="av" x:Class="EksamensProjekt.HistoryView"
         Title="History View" Height="1080" Width="1920" ResizeMode="NoResize" WindowStartupLocation="CenterScreen">
