+++ EksamensProjekt/Views/LoginView.xaml
@@ -4,7 +4,7 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:EksamensProjekt"
-        xmlns:util="clr-namespace:EksamensProjekt.Utilities"
+        xmlns:util="clr-namespace:EksamensProjekt.ViewsUtilities"
         mc:Ignorable="d"
         Title="MainWindow" Height="646" Width="999"
         ResizeMode="NoResize"
