+++ EksamensProjekt/Views/TenancyUploadView.xaml
@@ -4,7 +4,7 @@
          xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
          xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
          xmlns:local="clr-namespace:EksamensProjekt.Models"
-         xmlns:util="clr-namespace:EksamensProjekt.Utilities"
+         xmlns:util="clr-namespace:EksamensProjekt.ViewsUtilities"
          xmlns:sys="clr-namespace:System;assembly=mscorlib"
          mc:Ignorable="d"
          Title="Upload Tenancy"
